+++ CHANGES.rst
@@ -5,7 +5,8 @@ Changelog
 0.1.6 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Fix generic serializer/deserializer.
+  [cekk]
 
 
 0.1.5 (2020-10-20)